+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.4](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.8.3...@coveo/headless@2.8.4) (2023-02-08)
+
+### Bug Fixes
+
+- **headless:** provide proper instant results query for analytics ([#2719](https://github.com/coveo/ui-kit/issues/2719)) ([d40668f](https://github.com/coveo/ui-kit/commit/d40668fb1cedb5805001aaa331a0ac1991885cc5))
+
+## [2.8.3](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.8.2...@coveo/headless@2.8.3) (2023-02-06)
+
+### Bug Fixes
+
+- **atomic,headless:** update history with replaceState (before search) or pushState (after search) & validate tab values ([#2704](https://github.com/coveo/ui-kit/issues/2704)) ([5ae1115](https://github.com/coveo/ui-kit/commit/5ae1115abfc60fe198f55c8474786373d5e13f92))
+
 ## [2.8.2](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.8.1...@coveo/headless@2.8.2) (2023-02-06)
 
 **Note:** Version bump only for package @coveo/headless